Split artifact tests into separate files
